Working check of spectra and changed Well class (a bit)
