--- mathtools.py
@@ -1,16 +1,0 @@
-import numpy as np
-from scipy import sparse
-
-
-# baseline calculation by Asymmetric Least Squares Smoothing
-def baseline_als(y, lam=1e+7, p=0.001, niter=10):
-    L = len(y)
-    D = sparse.csc_matrix(np.diff(np.eye(L), 2))
-    w = np.ones(L)
-    for i in range(niter):
-        W = sparse.spdiags(w, 0, L, L)
-        Z = W + lam * D.dot(D.transpose())
-        z = sparse.linalg.spsolve(Z, w * y)
-        w = p * (y > z) + (1 - p) * (y < z)
-
-    return z
